# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AIGovOps REPLAY API - cryptographically signed AI interaction receipts
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /policies:
    get:
      operationId: listPolicies
      tags: [policies]
      summary: List all policy rules
      responses:
        "200":
          description: List of policies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyList"
    post:
      operationId: createPolicy
      tags: [policies]
      summary: Create a new policy rule
      description: Requires admin privilege (caller's user ID must be in the ADMIN_USER_IDS server environment variable).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyBody"
      responses:
        "201":
          description: Created policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "403":
          description: Forbidden — admin privilege required
  /policies/{id}:
    get:
      operationId: getPolicy
      tags: [policies]
      summary: Get a single policy rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
    patch:
      operationId: updatePolicy
      tags: [policies]
      summary: Update a policy rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePolicyBody"
      responses:
        "200":
          description: Updated policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
    delete:
      operationId: deletePolicy
      tags: [policies]
      summary: Delete a policy rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token passed as a Bearer token in the Authorization header (used by mobile clients)
...
  schemas:
...
    Policy:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        rule:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        enabled:
          type: boolean
        violationCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - rule
        - severity
        - enabled
        - violationCount
        - createdAt
        - updatedAt
    PolicyList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Policy"
        total:
          type: integer
      required:
        - items
        - total
    CreatePolicyBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        rule:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        enabled:
          type: boolean
      required:
        - name
        - description
        - rule
        - severity
    UpdatePolicyBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        rule:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        enabled:
          type: boolean
...